backend/app/core/crypto.py
import base64
import hashlib

from cryptography.fernet import Fernet

from app.core.config import get_settings


def _build_fernet() -> Fernet:
    settings = get_settings()
    digest = hashlib.sha256(settings.secret_key.encode("utf-8")).digest()
    key = base64.urlsafe_b64encode(digest)
    return Fernet(key)


def encrypt_text(raw: str) -> str:
    token = _build_fernet().encrypt(raw.encode("utf-8"))
    return token.decode("utf-8")


def decrypt_text(ciphertext: str) -> str:
    plain = _build_fernet().decrypt(ciphertext.encode("utf-8"))
    return plain.decode("utf-8")
