backend/pyproject.toml
[project]
name = "qrgift-backend"
version = "0.1.0"
description = "QRGift backend service"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
  "fastapi>=0.116.0",
  "uvicorn[standard]>=0.35.0",
  "sqlalchemy>=2.0.43",
  "alembic>=1.16.5",
  "pydantic-settings>=2.11.0",
  "python-jose[cryptography]>=3.5.0",
  "python-multipart>=0.0.20",
  "qrcode[pil]>=8.2",
  "opencv-python-headless>=4.10.0.84",
  "minio>=7.2.18",
  "oss2>=2.19.1"
]

[project.optional-dependencies]
dev = [
  "pytest>=8.4.2",
  "ruff>=0.13.0",
  "black>=25.9.0",
  "mypy>=1.18.2"
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["app"]

[tool.black]
line-length = 100
target-version = ["py311"]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "B", "UP"]

[tool.pytest.ini_options]
pythonpath = ["."]
